requirements.txt
accelerate==1.10.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
aistudio_sdk==0.3.5
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.11
attrs==25.3.0
backoff==2.2.1
bce-python-sdk==0.9.42
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2025.7.14
cfgv==3.4.0
chardet==5.2.0
charset-normalizer==3.4.2
chromadb==1.0.15
click==8.2.1
cohere==5.16.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
dill==0.4.0
distlib==0.4.0
distro==1.9.0
dnspython==2.7.0
docling==2.47.1
docling-core==2.45.0
docling-ibm-models==3.9.0
docling-parse==4.2.3
durationpy==0.10
easyocr==1.7.2
einops==0.8.1
et_xmlfile==2.0.0
faiss-cpu==1.11.0.post1
fastapi==0.116.1
fastavro==1.11.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.7.0
future==1.0.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hf-xet==1.1.5
hpack==4.1.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.33.4
humanfriendly==10.0
hyperframe==6.1.0
identify==2.6.13
idna==3.10
imageio==2.37.0
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
isort==6.0.1
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
jsonlines==3.1.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.9
kubernetes==33.1.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.69
langchain-text-splitters==0.3.8
langsmith==0.4.8
latex2mathml==3.78.0
lazy_loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
markdownify==1.2.0
marko==2.2.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.5
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.1.0
modelscope==1.29.1
mpire==2.10.2
mpmath==1.3.0
multidict==6.6.3
multiprocess==0.70.18
mypy_extensions==1.1.0
narwhals==2.1.2
networkx==3.5
ninja==1.13.0
nodeenv==1.9.1
numpy==2.2.6
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==12.575.51
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.8.90
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.97.0
opencv-contrib-python==4.10.0.84
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
opentelemetry-api==1.35.0
opentelemetry-exporter-otlp-proto-common==1.35.0
opentelemetry-exporter-otlp-proto-grpc==1.35.0
opentelemetry-proto==1.35.0
opentelemetry-sdk==1.35.0
opentelemetry-semantic-conventions==0.56b0
opt-einsum==3.3.0
orjson==3.11.0
overrides==7.7.0
packaging==25.0
paddleocr==3.2.0
paddlepaddle==3.1.1
paddlex==3.2.0
pandas==2.3.1
pillow==10.4.0
platformdirs==4.3.8
plotly==6.3.0
pluggy==1.6.0
portalocker==3.2.0
posthog==5.4.0
pre_commit==4.3.0
prettytable==3.16.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.1
pyclipper==1.3.0.post6
pycryptodome==3.23.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
pylatexenc==2.10
pylint==3.3.7
pymongo==4.13.2
PyMuPDF==1.26.3
pynvml==12.0.0
pyparsing==3.2.3
pypdfium2==4.30.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pysqlite3-binary==0.5.4
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pyTigerGraph==1.9.0
pytz==2025.2
PyYAML==6.0.2
qdrant-client==1.15.0
redis==6.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.0.0
rpds-py==0.26.0
rsa==4.9.1
rtree==1.4.1
ruamel.yaml==0.18.15
ruamel.yaml.clib==0.2.12
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.1
scipy==1.16.0
semchunk==2.2.2
sentence-transformers==5.0.0
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
SQLite3-0611==0.0.1
starlette==0.47.2
surya-ocr==0.15.4
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tifffile==2025.6.11
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
transformers==4.53.2
triton==3.3.1
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
virtualenv==20.34.0
watchfiles==1.1.0
wcwidth==0.2.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
xlsxwriter==3.2.5
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
